+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-### Added
+### [1.2.3] -2019-01-16
 - format should write zero value of boolean in settings, by [@bradrydzewski](https://github.com/bradrydzewski).
 - Support for windows server 1903, by [@bradrydzewski](https://github.com/bradrydzewski).
+- Support to convert legacy pipelines to kubernetes pipelines, by [@bradrydzewski](https://github.com/bradrydzewski).
 
 ## [1.2.2] - 2019-07-29
 ### Added